@@ -1,46 +1,30 @@
 # frozen_string_literal: true
 
-# == Schema Information
-#
-# Table name: teams
-#
-#  id                     :integer          not null, primary key
-#  name                   :string
-#  general_info           :text
-#  created_at             :datetime         not null
-#  updated_at             :datetime         not null
-#  logo_file_name         :string
-#  logo_content_type      :string
-#  logo_file_size         :integer
-#  logo_updated_at        :datetime
-#  slug                   :string
-#  preferences            :json
-#  slack_team_id          :string
-#  slack_bot_access_token :string
-#  channel_id             :string
-#
-
-
-class Team < ActiveRecord::Base
+class Team < ApplicationRecord
   extend FriendlyId
   friendly_id :slug_candidates, use: :slugged
   after_create :setup_team
-  after_validation :logo_reverted?
-  has_attached_file :logo, styles: { thumb: '320x120' }
+
   validates :name, presence: true
   validates :slug, presence: true
-  validates_attachment :logo, content_type: { content_type: %w[image/jpeg image/png] }
-  validates_with AttachmentSizeValidator, attributes: :logo, less_than: 10.megabytes
-  process_in_background :logo
+  validates :logo, file_content_type: {
+      allow: %w[image/jpeg image/png],
+      if: -> { logo.attached? }
+  }
 
   has_secure_token :rss_token
-  has_many :memberships, class_name: 'TeamMember', foreign_key: :team_id
+
+  has_many :memberships, class_name: "TeamMember", foreign_key: :team_id
   has_many :users, through: :memberships
-  has_many :balances
-  has_many :goals, through: :balances
-  has_many :transactions
+  has_one :active_kudos_meter, class_name: "KudosMeter"
+  has_many :kudos_meters
+  has_many :goals, through: :kudos_meters
+  has_many :posts
   has_many :guidelines
-  has_many :likes, class_name: 'Vote'
+  has_many :team_invites
+  has_many :likes, class_name: "Vote"
+
+  has_one_attached :logo
 
   typed_store :preferences, coder: PreferencesCoder do |p|
     p.string :primary_color, default: nil
@@ -52,16 +36,22 @@ class Team < ActiveRecord::Base
 
   def name_and_sequence
     slug = name.parameterize
-    sequence = Team.where('slug like ?', "#{slug}-%").count + 2
-    "#{slug}-#{sequence}"
+    sequence = Team.where("slug like ?", "#{slug}%")
+
+    return slug if sequence.empty?
+
+    "#{slug}-#{sequence.count + 1}"
   end
 
-  def add_member(user, admin = false)
-    TeamMember.create(user: user, team: self, admin: admin)
+  def add_member(user, role = 'member')
+    TeamMember.create(user: user, team: self, role: role)
   end
 
   def remove_member(user)
-    TeamMember.find_by_user_id_and_team_id(user.id, id).delete
+    transaction do
+      TeamMember.find_by_user_id_and_team_id(user.id, id).delete
+      TeamInvite.where(email: user.email, team_id: id).delete_all
+    end
   end
 
   def member?(user)
@@ -74,12 +64,16 @@ class Team < ActiveRecord::Base
 
   def manageable_members(current_user)
     memberships.joins(:user)
-               .where('users.company_user = false')
-               .where("users.id != #{current_user.id}")
+     .where("users.company_user = false")
+     .where("users.id != #{current_user.id}")
   end
 
   def current_goals
-    goals.where(balance: Balance.current(id))
+    goals.where(kudos_meter: Team.find(id).active_kudos_meter)
+  end
+
+  def achieved_goals
+    goals.where(kudos_meter: Team.find(id).active_kudos_meter).where.not(achieved_on: nil)
   end
 
   def real_users
@@ -89,15 +83,13 @@ class Team < ActiveRecord::Base
   private
 
   def setup_team
-    # Create balances and goals
-    balance = Balance.create(name: 'My first balance', current: true,
-                             team_id: id)
-    Goal.create(name: 'First goal', amount: 500, balance_id: balance.id)
-    Goal.create(name: 'Second goal', amount: 1000, balance_id: balance.id)
-    Goal.create(name: 'Third goal', amount: 1500, balance_id: balance.id)
+    kudos_meter = KudosMeter.create(name: "My first kudos_meter", team_id: id)
+    Goal.create(name: "First goal", amount: 500, kudos_meter_id: kudos_meter.id)
+    Goal.create(name: "Second goal", amount: 1000, kudos_meter_id: kudos_meter.id)
+    Goal.create(name: "Third goal", amount: 1500, kudos_meter_id: kudos_meter.id)
 
     # Create company user
-    user = User.new(name: name, company_user: true)
+    user = User.new(name: name, company_user: true, virtual_user: true)
     user.save(validate: false)
     add_member(user)
   end
