@@ -1,19 +1,6 @@
-# == Schema Information
-#
-# Table name: team_members
-#
-#  id             :integer          not null, primary key
-#  team_id        :integer
-#  user_id        :integer
-#  admin          :boolean
-#  created_at     :datetime         not null
-#  updated_at     :datetime         not null
-#  slack_id       :string
-#  slack_username :string
-#  slack_name     :string
-#
+# frozen_string_literal: true
 
-class TeamMember < ActiveRecord::Base
+class TeamMember < ApplicationRecord
   belongs_to :team
   belongs_to :user
 end
