@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class GoalDecorator < Draper::Decorator
   delegate_all
 
@@ -6,7 +8,6 @@ class GoalDecorator < Draper::Decorator
   end
 
   def kudos
-    h.number_to_kudos(object.amount)
+    ApplicationController.helpers.number_to_kudos(object.amount)
   end
-
 end
