@@ -1,21 +1,23 @@
+# frozen_string_literal: true
+
 class KudosDeviseMailer < Devise::Mailer
   helper :application # gives access to all helpers defined within `application_helper`.
   include Devise::Controllers::UrlHelpers # Optional. eg. `confirmation_url`
-  default template_path: 'devise/mailer' # to make sure that your mailer uses the devise views
-  layout 'mailer'
-  default from: ENV['MAIL_USERNAME']
+  default template_path: "devise/mailer" # to make sure that your mailer uses the devise views
+  layout "mailer"
+  default from: ENV["MAIL_USERNAME"]
 
-  def confirmation_instructions(record, token, opts={})
+  def confirmation_instructions(record, token, opts = {})
     add_logo_attachment
     super
   end
 
-  def reset_password_instructions(record, token, opts={})
+  def reset_password_instructions(record, token, opts = {})
     add_logo_attachment
     super
   end
 
-  def unlock_instructions(record, token, opts={})
+  def unlock_instructions(record, token, opts = {})
     add_logo_attachment
     super
   end
@@ -23,6 +25,6 @@ class KudosDeviseMailer < Devise::Mailer
   private
 
   def add_logo_attachment
-    attachments.inline['logo.png'] = File.read("#{Rails.root}/app/assets/images/kudo-o-matic-white-mail.png")
+    attachments.inline["logo.png"] = File.read("#{Rails.root}/app/assets/images/kudo-o-matic-white-mail.png")
   end
 end
