@@ -1,10 +1,6 @@
+# frozen_string_literal: true
+
 class UserMailer < ApplicationMailer
-  def self.new_user(user)
-    return if Rails.env == 'test' || ENV['MAIL_USERNAME'].blank? || user.email.blank?
-
-    suppress(Exception) {welcome_email(user).deliver_later}
-  end
-
   def invite_email(email, team)
     @team = team
 
@@ -14,19 +10,6 @@ class UserMailer < ApplicationMailer
   def welcome_email(user)
     @user = user
 
-    mail(to: user.email, subject: 'Welcome to the ₭udo-o-Matic!')
-  end
-
-  def export_start_email(user)
-    @user = user
-
-    mail(to: user.email, subject: 'Started data export!')
-  end
-
-  def export_done_email(user, export)
-    @user = user
-
-    @export = export
-    mail(to: user.email, subject: 'Data export finished!')
+    mail(to: user.email, subject: "Welcome to the ₭udo-o-Matic!")
   end
 end
