@@ -1,11 +1,13 @@
+# frozen_string_literal: true
+
 class ExportService
   include Singleton
 
   def start_new_export(user, format)
-    Delayed::Job.enqueue Export::CreateExportJob.new(user, format)
+    Export::CreateExportJob.new(user, format)
   end
 
   def delete_expired_exports
-    Delayed::Job.enqueue Export::MaintainExportsJob.new
+    Export::MaintainExportsJob.new
   end
 end
