@@ -1,20 +1,10 @@
+# frozen_string_literal: true
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
 
-  # paperclip S3
-  config.paperclip_defaults = {
-      storage: :s3,
-      s3_protocol: :https,
-      bucket: ENV["AWS_S3_BUCKET"],
-      s3_credentials: {
-          s3_host_name: ENV["AWS_S3_HOST_NAME"],
-          s3_region: ENV["AWS_S3_REGION"],
-          access_key_id: ENV["AWS_ACCESS_KEY_ID"],
-          secret_access_key: ENV["AWS_SECRET_ACCESS_KEY"],
-          path: "image/:id/:filename",
-          url: ':s3_domain_url'
-      }
-  }
+  # Store files on amazon.
+  config.active_storage.service = :local
 
   # In the development environment your application's code is reloaded on
   # every request. This slows down response time but is perfect for development
@@ -28,12 +18,12 @@ Rails.application.configure do
   config.consider_all_requests_local = true
 
   # Enable/disable caching. By default caching is disabled.
-  if Rails.root.join('tmp/caching-dev.txt').exist?
+  if Rails.root.join("tmp/caching-dev.txt").exist?
     config.action_controller.perform_caching = true
 
     config.cache_store = :memory_store
     config.public_file_server.headers = {
-      'Cache-Control' => 'public, max-age=172800'
+      "Cache-Control" => "public, max-age=172800"
     }
   else
     config.action_controller.perform_caching = false
@@ -52,7 +42,7 @@ Rails.application.configure do
   # SMTP settings email
   config.action_mailer.smtp_settings = {
       address: "localhost",
-      port: 1025,
+      port: 1025
   }
 
   # config.action_mailer.default_url_options = {host: 'localhost', port: 3000}
