@@ -1,17 +1,7 @@
-# == Schema Information
-#
-# Table name: activities
-#
-#  id               :integer          not null, primary key
-#  name             :text
-#  suggested_amount :integer
-#  created_at       :datetime         not null
-#  updated_at       :datetime         not null
-#
+# frozen_string_literal: true
 
 FactoryBot.define do
   factory :activity do
-    name "getting me coffee"
+    name { "getting me coffee" }
   end
 end
-
