@@ -2,8 +2,8 @@ require "rails_helper"
 
 RSpec.describe KudosDeviseMailer, type: :mailer do
   context 'confirmation instruction' do
-    let(:user) { User.create name: 'John Doe', email:'johndoe@example.com' }
-    let(:mail) {described_class.confirmation_instructions(user, "faketoken", {})}
+    let(:user) { User.create name: 'John Doe', email: 'johndoe@example.com' }
+    let(:mail) { described_class.confirmation_instructions(user, "faketoken", {}) }
 
     it 'renders the subject' do
       expect(mail.subject).to eq('Confirmation instructions')
@@ -14,7 +14,7 @@ RSpec.describe KudosDeviseMailer, type: :mailer do
     end
 
     it 'sends the email' do
-      expect {mail.deliver_now}.to change {ActionMailer::Base.deliveries.count}.from(0).to(1)
+      expect { mail.deliver_now}.to change {ActionMailer::Base.deliveries.count }.from(0).to(1)
     end
 
     it 'logo attachment is added' do
@@ -24,7 +24,7 @@ RSpec.describe KudosDeviseMailer, type: :mailer do
 
   context 'reset password instruction' do
     let(:user) { User.create name: 'John Doe', email:'johndoe@example.com' }
-    let(:mail) {described_class.reset_password_instructions(user, "faketoken", {})}
+    let(:mail) { described_class.reset_password_instructions(user, "faketoken", {}) }
 
     it 'renders the subject' do
       expect(mail.subject).to eq('Reset password instructions')
@@ -35,7 +35,7 @@ RSpec.describe KudosDeviseMailer, type: :mailer do
     end
 
     it 'sends the email' do
-      expect {mail.deliver_now}.to change {ActionMailer::Base.deliveries.count}.from(0).to(1)
+      expect { mail.deliver_now }.to change { ActionMailer::Base.deliveries.count }.from(0).to(1)
     end
 
     it 'logo attachment is added' do
@@ -45,7 +45,7 @@ RSpec.describe KudosDeviseMailer, type: :mailer do
 
   context 'unlock instruction' do
     let(:user) { User.create name: 'John Doe', email:'johndoe@example.com' }
-    let(:mail) {described_class.unlock_instructions(user, "faketoken", {})}
+    let(:mail) { described_class.unlock_instructions(user, "faketoken", {}) }
 
     it 'renders the subject' do
       expect(mail.subject).to eq('Unlock instructions')
@@ -56,7 +56,7 @@ RSpec.describe KudosDeviseMailer, type: :mailer do
     end
 
     it 'sends the email' do
-      expect {mail.deliver_now}.to change {ActionMailer::Base.deliveries.count}.from(0).to(1)
+      expect { mail.deliver_now }.to change { ActionMailer::Base.deliveries.count }.from(0).to(1)
     end
 
     it 'logo attachment is added' do
